refactor: 주행 히스토리 및 주행상세 리팩토링
