+++ src/theme/colors.ts
@@ -1,0 +1,7 @@
+export const colors = {
+  primary: '#4945FF',
+  neutralLight: '#F2F2FF',
+  neutralDark: '#666666',
+  timelineLine: '#D0D0D0',
+  background: '#FFFFFF',
+};
